rename global.css.js to global-styles.js
